+++ ReleaseNotes.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## 0.2.3
+
+- *resource stream* feature
+- `aws`, `sql` blob storage options; `quest-py` extras
+- fixed historian storage memory leak, storages clearing properly
+- added SIGINT handler
+- added workflow metrics
+
 ## 0.2.2
 
 - `alias` feature